+++ src-old/hooks/use-window-breakpoints/types.ts
@@ -1,8 +1,0 @@
-export interface UseWindowBreakpoints {
-    width: number;
-    xs: boolean;
-    sm: boolean;
-    md: boolean;
-    lg: boolean;
-    xl: boolean;
-}