+++ src/index.ts
@@ -28,6 +28,7 @@ export * from "./widgets/multi-select-field/multi-select-field";
 
 export * from "./utils/use-query-params/use-query-params";
 export * from "./utils/use-countdown/use-countdown";
+export * from "./utils/use-window-breakpoints/use-window-breakpoints";
 
 export * from "../src-old/form/file-picker";
 export * from "../src-old/form/file-field";
@@ -37,6 +38,5 @@ export * from "../src-old/form/time-picker";
 export * from "../src-old/form/time-field";
 
 export * from "../src-old/hooks/use-load-more";
-export * from "../src-old/hooks/use-window-breakpoints";
 
 export * from "../src-old/misc/debounce";